+++ providers/gcs/gcs_test.go
@@ -16,6 +16,7 @@ import (
 	"github.com/fullstorydev/emulators/storage/gcsemu"
 	"github.com/go-kit/log"
 	"github.com/prometheus/common/model"
+	"github.com/thanos-io/objstore/errutil"
 	"google.golang.org/api/option"
 )
 
@@ -66,11 +67,41 @@ func TestNewBucketWithConfig_ShouldCreateGRPC(t *testing.T) {
 	err = os.Setenv("STORAGE_EMULATOR_HOST_GRPC", svr.Addr)
 	testutil.Ok(t, err)
 
-	bkt, err := NewBucketWithConfig(context.Background(), log.NewNopLogger(), cfg, "test-bucket")
+	bkt, err := NewBucketWithConfig(context.Background(), log.NewNopLogger(), cfg, "test-bucket", nil)
 	testutil.Ok(t, err)
 
 	// Check if the bucket is created.
 	testutil.Assert(t, bkt != nil, "expected bucket to be created")
+}
+
+func TestParseConfig_ChunkSize(t *testing.T) {
+	for _, tc := range []struct {
+		name       string
+		input      string
+		assertions func(cfg Config)
+	}{
+		{
+			name:  "DefaultConfig",
+			input: `bucket: abcd`,
+			assertions: func(cfg Config) {
+				testutil.Equals(t, cfg.ChunkSizeBytes, 0)
+			},
+		},
+		{
+			name: "CustomConfig",
+			input: `bucket: abcd
+chunk_size_bytes: 1024`,
+			assertions: func(cfg Config) {
+				testutil.Equals(t, cfg.ChunkSizeBytes, 1024)
+			},
+		},
+	} {
+		t.Run(tc.name, func(t *testing.T) {
+			cfg, err := parseConfig([]byte(tc.input))
+			testutil.Ok(t, err)
+			tc.assertions(cfg)
+		})
+	}
 }
 
 func TestParseConfig_HTTPConfig(t *testing.T) {
@@ -126,4 +157,24 @@ http_config:
 			tc.assertions(cfg)
 		})
 	}
+}
+
+func TestNewBucketWithErrorRoundTripper(t *testing.T) {
+	cfg := Config{
+		Bucket:         "test-bucket",
+		ServiceAccount: "",
+		UseGRPC:        false,
+		noAuth:         true,
+	}
+	svr, err := gcsemu.NewServer("127.0.0.1:0", gcsemu.Options{})
+	testutil.Ok(t, err)
+	defer svr.Close()
+	err = os.Setenv("STORAGE_EMULATOR_HOST", svr.Addr)
+	testutil.Ok(t, err)
+
+	bkt, err := NewBucketWithConfig(context.Background(), log.NewNopLogger(), cfg, "test-bucket", errutil.WrapWithErrRoundtripper)
+	testutil.Ok(t, err)
+	_, err = bkt.Get(context.Background(), "test-bucket")
+	testutil.NotOk(t, err)
+	testutil.Assert(t, errutil.IsMockedError(err), "Expected RoundTripper error, got: %v", err)
 }