+++ providers/oci/oci.go
@@ -58,13 +58,14 @@ type HTTPConfig struct {
 	ResponseHeaderTimeout model.Duration `yaml:"response_header_timeout"`
 	InsecureSkipVerify    bool           `yaml:"insecure_skip_verify"`
 
-	TLSHandshakeTimeout   model.Duration `yaml:"tls_handshake_timeout"`
-	ExpectContinueTimeout model.Duration `yaml:"expect_continue_timeout"`
-	MaxIdleConns          int            `yaml:"max_idle_conns"`
-	MaxIdleConnsPerHost   int            `yaml:"max_idle_conns_per_host"`
-	MaxConnsPerHost       int            `yaml:"max_conns_per_host"`
-	DisableCompression    bool           `yaml:"disable_compression"`
-	ClientTimeout         time.Duration  `yaml:"client_timeout"`
+	TLSHandshakeTimeout   model.Duration    `yaml:"tls_handshake_timeout"`
+	ExpectContinueTimeout model.Duration    `yaml:"expect_continue_timeout"`
+	MaxIdleConns          int               `yaml:"max_idle_conns"`
+	MaxIdleConnsPerHost   int               `yaml:"max_idle_conns_per_host"`
+	MaxConnsPerHost       int               `yaml:"max_conns_per_host"`
+	DisableCompression    bool              `yaml:"disable_compression"`
+	ClientTimeout         time.Duration     `yaml:"client_timeout"`
+	Transport             http.RoundTripper `yaml:"-"`
 }
 
 // Config stores the configuration for oci bucket.
@@ -133,7 +134,12 @@ func (b *Bucket) Get(ctx context.Context, name string) (io.ReadCloser, error) {
 	if err != nil {
 		return nil, err
 	}
-	return response.Content, nil
+	return objstore.ObjectSizerReadCloser{
+		ReadCloser: response.Content,
+		Size: func() (int64, error) {
+			return *response.ContentLength, nil
+		},
+	}, nil
 }
 
 // GetRange returns a new range reader for the given object name and range.
@@ -163,7 +169,11 @@ func (b *Bucket) GetRange(ctx context.Context, name string, offset, length int64
 	if err != nil {
 		return nil, err
 	}
-	return response.Content, nil
+	return objstore.ObjectSizerReadCloser{ReadCloser: response.Content,
+		Size: func() (int64, error) {
+			return *response.ContentLength, nil
+		},
+	}, nil
 }
 
 // Upload the contents of the reader as an object into the bucket.
@@ -288,7 +298,7 @@ func (b *Bucket) deleteBucket(ctx context.Context) (err error) {
 }
 
 // NewBucket returns a new Bucket using the provided oci config values.
-func NewBucket(logger log.Logger, ociConfig []byte) (*Bucket, error) {
+func NewBucket(logger log.Logger, ociConfig []byte, wrapRoundtripper func(http.RoundTripper) http.RoundTripper) (*Bucket, error) {
 	level.Debug(logger).Log("msg", "creating new oci bucket connection")
 	var config = DefaultConfig
 	var configurationProvider common.ConfigurationProvider
@@ -334,9 +344,16 @@ func NewBucket(logger log.Logger, ociConfig []byte) (*Bucket, error) {
 	if err != nil {
 		return nil, errors.Wrapf(err, "unable to create ObjectStorage client with the given oci configurations")
 	}
-
+	var rt http.RoundTripper
+	rt = CustomTransport(config)
+	if config.HTTPConfig.Transport != nil {
+		rt = config.HTTPConfig.Transport
+	}
+	if wrapRoundtripper != nil {
+		rt = wrapRoundtripper(rt)
+	}
 	httpClient := http.Client{
-		Transport: CustomTransport(config),
+		Transport: rt,
 		Timeout:   config.HTTPConfig.ClientTimeout,
 	}
 	client.HTTPClient = &httpClient
@@ -375,7 +392,7 @@ func NewTestBucket(t testing.TB) (objstore.Bucket, func(), error) {
 		return nil, nil, err
 	}
 
-	bkt, err := NewBucket(log.NewNopLogger(), ociConfig)
+	bkt, err := NewBucket(log.NewNopLogger(), ociConfig, nil)
 	if err != nil {
 		return nil, nil, err
 	}