+++ providers/oss/oss.go
@@ -16,12 +16,14 @@ import (
 	"testing"
 	"time"
 
+	"github.com/aliyun/aliyun-oss-go-sdk/oss"
 	alioss "github.com/aliyun/aliyun-oss-go-sdk/oss"
 	"github.com/go-kit/log"
 	"github.com/pkg/errors"
 	"gopkg.in/yaml.v2"
 
 	"github.com/thanos-io/objstore/clientutil"
+	"github.com/thanos-io/objstore/exthttp"
 
 	"github.com/thanos-io/objstore"
 )
@@ -158,22 +160,32 @@ func (b *Bucket) Attributes(ctx context.Context, name string) (objstore.ObjectAt
 }
 
 // NewBucket returns a new Bucket using the provided oss config values.
-func NewBucket(logger log.Logger, conf []byte, component string) (*Bucket, error) {
+func NewBucket(logger log.Logger, conf []byte, component string, wrapRoundtripper func(http.RoundTripper) http.RoundTripper) (*Bucket, error) {
 	var config Config
 	if err := yaml.Unmarshal(conf, &config); err != nil {
 		return nil, errors.Wrap(err, "parse aliyun oss config file failed")
 	}
-
-	return NewBucketWithConfig(logger, config, component)
+	return NewBucketWithConfig(logger, config, component, wrapRoundtripper)
 }
 
 // NewBucketWithConfig returns a new Bucket using the provided oss config struct.
-func NewBucketWithConfig(logger log.Logger, config Config, component string) (*Bucket, error) {
+func NewBucketWithConfig(logger log.Logger, config Config, component string, wrapRoundtripper func(http.RoundTripper) http.RoundTripper) (*Bucket, error) {
 	if err := validate(config); err != nil {
 		return nil, err
 	}
-
-	client, err := alioss.New(config.Endpoint, config.AccessKeyID, config.AccessKeySecret)
+	var clientOptions []alioss.ClientOption
+	if wrapRoundtripper != nil {
+		rt, err := exthttp.DefaultTransport(exthttp.DefaultHTTPConfig)
+		if err != nil {
+			return nil, err
+		}
+		clientOptions = append(clientOptions, func(client *alioss.Client) {
+			client.HTTPClient = &http.Client{
+				Transport: wrapRoundtripper(rt),
+			}
+		})
+	}
+	client, err := alioss.New(config.Endpoint, config.AccessKeyID, config.AccessKeySecret, clientOptions...)
 	if err != nil {
 		return nil, errors.Wrap(err, "create aliyun oss client failed")
 	}
@@ -274,7 +286,7 @@ func NewTestBucketFromConfig(t testing.TB, c Config, reuseBucket bool) (objstore
 		return nil, nil, err
 	}
 
-	b, err := NewBucket(log.NewNopLogger(), bc, "thanos-aliyun-oss-test")
+	b, err := NewBucket(log.NewNopLogger(), bc, "thanos-aliyun-oss-test", nil)
 	if err != nil {
 		return nil, nil, err
 	}
@@ -338,12 +350,22 @@ func (b *Bucket) getRange(_ context.Context, name string, off, length int64) (io
 		opts = append(opts, opt)
 	}
 
-	resp, err := b.bucket.GetObject(name, opts...)
+	resp, err := b.bucket.DoGetObject(&oss.GetObjectRequest{ObjectKey: name}, opts)
 	if err != nil {
 		return nil, err
 	}
 
-	return resp, nil
+	size, err := clientutil.ParseContentLength(resp.Response.Headers)
+	if err == nil {
+		return objstore.ObjectSizerReadCloser{
+			ReadCloser: resp.Response,
+			Size: func() (int64, error) {
+				return size, nil
+			},
+		}, nil
+	}
+
+	return resp.Response, nil
 }
 
 // Get returns a reader for the given object name.